Begin converting the get words function from python implementation into rust state machine iterator

+++ src/consts.rs
@@ -1,1 +1,1 @@
-pub const OPEN: u8 = 255;
+pub const OPEN: usize = 255;
