+++ src/main.rs
@@ -5,7 +5,7 @@ mod trie;
 mod util;
 
 fn main() {
-    let grid = grid::Grid::<u8, 9, 3, 3>::new();
+    let grid = grid::Grid::<usize, 9, 3, 3>::new();
     let char_grid = grid.try_map(util::int_to_char).unwrap();
     println!("{}", char_grid);
