+++ src/util.rs
@@ -1,8 +1,8 @@
 use crate::consts::OPEN;
 
-pub fn char_to_int(c: char) -> Result<u8, String> {
+pub fn char_to_int(c: char) -> Result<usize, String> {
     if c.is_ascii_alphabetic() {
-        Ok(c.to_ascii_lowercase() as u8 - b'a')
+        Ok(c.to_ascii_lowercase() as usize - b'a' as usize)
     } else if c == '_' {
         Ok(OPEN)
     } else {
@@ -10,9 +10,9 @@ pub fn char_to_int(c: char) -> Result<u8, String> {
     }
 }
 
-pub fn int_to_char(n: &u8) -> Result<char, String> {
+pub fn int_to_char(n: &usize) -> Result<char, String> {
     if *n < 26 {
-        Ok((b'a' + n) as char)
+        Ok((b'a' as usize + n) as u8 as char)
     } else if *n == OPEN {
         Ok('_')
     } else {
@@ -20,10 +20,10 @@ pub fn int_to_char(n: &u8) -> Result<char, String> {
     }
 }
 
-pub fn string_to_ints(s: &str) -> Result<Vec<u8>, String> {
+pub fn string_to_ints(s: &str) -> Result<Vec<usize>, String> {
     s.chars().map(char_to_int).collect()
 }
 
-fn ints_to_string(ints: Vec<u8>) -> Result<String, String> {
+fn ints_to_string(ints: Vec<usize>) -> Result<String, String> {
     ints.iter().map(int_to_char).collect()
 }
